Added note about fork status.
This isn't meant to be the future home of the project but I'd feel bad about having a fix available and not allowing others to use it in their own docker setups. So for now, this repo exists with a docker image being generated automatically thanks to a new github action.

If any API changes occur in the future and the upstream owner doesn't return, I'm okay with making patches and pushing image updates.